update file structure.
